rm unused code and shrink codes into seperate files
