+++ ODIN_II/SRC/include/config_t.h
@@ -16,6 +16,7 @@ struct config_t
 	bool output_netlist_graphs; // switch that outputs netlist graphs per node for use with GraphViz tools
 	bool print_parse_tokens; // switch that controls whether or not each token is printed during parsing
 	bool output_preproc_source; // TODO: unused
+	bool ga_adder; // enable genetic algorithm for adder configuration
 	
 	int min_hard_multiplier; // threshold from hard to soft logic
 	int mult_padding; // setting how multipliers are padded to fit fixed size
