+++ ODIN_II/SRC/include/netlist_utils.h
@@ -13,9 +13,7 @@ npin_t* allocate_npin();
 npin_t *get_zero_pin(netlist_t *netlist);
 npin_t *get_pad_pin(netlist_t *netlist);
 npin_t *get_one_pin(netlist_t *netlist);
-//MEHRSHAD//
 npin_t *copy_npin(npin_t *copy_pin);
-//MEHRSHAD//
 npin_t* copy_input_npin(npin_t* copy_pin);
 npin_t* copy_output_npin(npin_t* copy_pin);
 nnet_t* allocate_nnet();
