+++ ODIN_II/SRC/include/odin_types.h
@@ -101,7 +101,7 @@ struct global_args_t
     argparse::ArgValue<bool> all_warnings;
     argparse::ArgValue<bool> show_help;
 
-	argparse::ArgValue<std::string> adder_def; //carry skip adder skip size
+	argparse::ArgValue<bool> ga_adder; // enable ga_adder
 	
     // defines if the first cin of an adder/subtractor is connected to a global gnd/vdd
     // or generated using a dummy adder with both inputs set to gnd/vdd
