+++ ODIN_II/SRC/include/partial_map.h
@@ -26,43 +26,6 @@ OTHER DEALINGS IN THE SOFTWARE.
 #ifndef PARTIAL_MAP_H
 #define PARTIAL_MAP_H
 
-//MEHRSHAD//
-
-#define DEFAULT  0
-#define CSLA     1
-#define BEC_CSLA 2
-
-#define CHILDREN_NUM 5
-
-struct generation_t {
-
-    short **chromosomes;
-    short *fittest;
-    int   *fitnesses;
-};
-
-struct adder_t {
-
-    short type = DEFAULT;
-
-    signal_list_t *input;
-    signal_list_t *output;
-
-    nnode_t *node;
-};
-
-void partial_map_adders_GA_top(netlist_t *netlist);
-void partial_map_adders(/*short traverse_number,*/ netlist_t *netlist);
-void destroy_adders();
-void destroy_adder_cloud (adder_t *adder);
-void recursive_save_pointers (adder_t *adder, nnode_t * node);
-npin_t** make_copy_of_pins (npin_t **copy, long copy_size);
-adder_t *create_empty_adder (adder_t *previous_adder);
-generation_t* create_generation (generation_t *previous_generation, int generation_counter);
-short *mutate (short *parent);
-
-//MEHRSHAD//
-
 // PROTOTYPES
 void partial_map_top(netlist_t *netlist);
 void instantiate_add_w_carry(short type, nnode_t *node, short mark, netlist_t *netlist);
